To place products in the database

+++ src/main/java/org/casjedcem/Farmshop/repository/ProducerRepository.java
@@ -12,9 +12,7 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface ProducerRepository extends JpaRepository<Producer, Long> {
-
-
+public interface ProducerRepository extends JpaRepository<Producer, Integer> {
 
 
 }
